requirements.txt
faker>=24.0.0
pymongo>=4.6.0
redis>=5.0.0
python-dotenv>=1.0.0
streamlit>=1.36.0
plotly>=5.20.0
pandas>=2.2.0
pydeck>=0.9.1
tenacity>=8.2.3
